django_test_proj/HillarDjangoREST/01/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
boltons==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
Django==1.11.17
django-filter==2.1.0
djangorestframework==3.6.4
httpie==1.0.2
idna==2.8
importlib-metadata==0.18
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==7.1.0
packaging==19.0
parso==0.4.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
py==1.8.0
Pygments==2.4.0
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
pytz==2019.1
requests==2.21.0
six==1.12.0
style==1.1.0
traitlets==4.3.2
update==0.0.1
urllib3==1.24.3
wcwidth==0.1.7
zipp==0.5.1
